Some small fixes

+++ libc_functions/ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: wkorande <wkorande@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/16 16:01:01 by wkorande          #+#    #+#             */
-/*   Updated: 2019/10/16 16:10:07 by wkorande         ###   ########.fr       */
+/*   Updated: 2019/10/16 19:23:55 by wkorande         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t			i;
+	int			i;
 
 	i = ft_strlen(s) - 1;
 	while (i >= 0)
